Refactoring UI file
