+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rask"
-version = "2.4.0"
+version = "2.4.1"
 edition = "2021"
 description = "An advanced CLI project planner with tags, priorities, dependencies, and multi-project support"
 license = "MIT"
