+++ src/ui/mod.rs
@@ -1,0 +1,16 @@
+pub mod dependencies;
+pub mod helpers;
+pub mod messages;
+pub mod progress;
+pub mod projects;
+pub mod roadmap;
+pub mod tasks;
+
+// Re-export commonly used functions
+pub use dependencies::*;
+pub use helpers::*;
+pub use messages::*;
+pub use progress::*;
+pub use projects::*;
+pub use roadmap::*;
+pub use tasks::*;
